nueva vista registro y control de errores en front end/ todavia no se esta enviando la data a back end
