+++ UGym/FrontEnd/Views/Account/Login.cshtml
@@ -28,7 +28,7 @@
                     <button type="submit" style="margin-bottom: 10px;">Ingresar</button>
                     <p class="mensaje">
                         ¿No tienes una cuenta?
-                        @Html.ActionLink("Crea tu Cuenta", "Register", "Account", new { area = "" }, new { @class = "waves-effect waves-dark" })
+                        @Html.ActionLink("Crea tu Cuenta", "Register", "FEAuthentication", new { area = "" }, new { @class = "waves-effect waves-dark" })
                     </p>
                 </form>
             </div>
