--- UGym/FrontEnd/Views/Authentication/Register.cshtml
@@ -1,7 +1,0 @@
-﻿@{
-    ViewBag.Title = "Registro";
-    Layout = "~/Views/Shared/_layout.cshtml";
-}
-
-<h1>Register</h1>
-
